memperbaiki hapu kelas

+++ src/main/java/com/example/absensireact/impl/KelasImpl.java
@@ -1,12 +1,15 @@
 package com.example.absensireact.impl;
 
 import com.example.absensireact.exception.NotFoundException;
+import com.example.absensireact.exception.RelationExistsException;
 import com.example.absensireact.model.Admin;
 import com.example.absensireact.model.Kelas;
 import com.example.absensireact.model.Organisasi;
+import com.example.absensireact.model.User;
 import com.example.absensireact.repository.AdminRepository;
 import com.example.absensireact.repository.KelasRepository;
 import com.example.absensireact.repository.OrganisasiRepository;
+import com.example.absensireact.repository.UserRepository;
 import com.example.absensireact.service.KelasService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -25,6 +28,9 @@ public class KelasImpl implements KelasService {
 
     @Autowired
     private OrganisasiRepository organisasiRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     @Override
     public List<Kelas>getAllKelas(){
@@ -73,7 +79,19 @@ public class KelasImpl implements KelasService {
 
     @Override
     public void deleteKelas(Long id) {
-        kelasRepository.deleteById(id);
+        if (checkIfHasRelations(id)) {
+            throw new RelationExistsException("Data ini sudah memiliki relasi dan tidak dapat dihapus");
+        } else {
+            kelasRepository.deleteById(id);
+        }
+    }
+
+
+    @Override
+    public boolean checkIfHasRelations(Long kelasId) {
+        // Use the repository method to check if there are users associated with the kelas
+        List<User> users = userRepository.findUsersByKelas(kelasId);
+        return !users.isEmpty(); // Return true if there are any users associated, otherwise false
     }
